src/data_generator/requirements.txt
requests==2.32.3
python-dotenv==1.0.0
azure-identity
azure-search-documents
azure-cosmos
openai>=1.30.0